requirements.txt
# Core ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
optuna>=3.4.0

# Data
pandas>=2.1.0
numpy>=1.25.0
ucimlrepo>=0.0.3

# Explainability
shap>=0.43.0
lime>=0.2.0.1
dice-ml>=0.11

# Fairness
fairlearn>=0.9.0
aif360>=0.6.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Config & Utils
pyyaml>=6.0.1
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
ruff>=0.1.6
